refactor: Split up datatable file
