+++ app/browser/dataset-preview.tsx
@@ -8,7 +8,7 @@ import { useRouter } from "next/router";
 import React, { ComponentProps, useEffect } from "react";
 import { UseQueryResponse } from "urql";
 
-import { CubeDataTablePreview } from "@/browse/datatable";
+import { CubeDataTablePreview } from "@/browse/cube-data-table-preview";
 import { useFootnotesStyles } from "@/components/chart-footnotes";
 import { DataDownloadMenu } from "@/components/data-download";
 import Flex from "@/components/flex";
