+++ app/components/chart-published.tsx
@@ -5,7 +5,7 @@ import clsx from "clsx";
 import { forwardRef, useCallback, useEffect, useMemo, useRef } from "react";
 import { useStore } from "zustand";
 
-import { ChartDataTablePreview } from "@/browse/datatable";
+import { ChartDataTablePreview } from "@/browse/chart-data-table-preview";
 import { extractChartConfigsComponentIds } from "@/charts/shared/chart-helpers";
 import { LoadingStateProvider } from "@/charts/shared/chart-loading-state";
 import { isUsingImputation } from "@/charts/shared/imputation";
